feat: 게시글 수정/삭제 권한 체크 추가
- 게시글 수정/삭제 시 작성자 검증 로직 추가
- JwtAuthGuard를 통한 인증 추가
- 권한이 없는 경우 403 Forbidden 응답 추가

+++ src/modules/board/board.service.ts
@@ -44,20 +44,26 @@ export class BoardService {
     return await this.boardRepository.save(board);
   }
 
-  async updateOne(id: number, data: UpdateBoardDto) {
+  async updateOne(userId: number, id: number, data: UpdateBoardDto) {
     const board = await this.getBoardById(id);
     if (!board) {
       throw new HttpException('NotFound', HttpStatus.NOT_FOUND);
+    }
+    if (board.userId !== userId) {
+      throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
     }
     return this.boardRepository.update(id, {
       ...data,
     });
   }
 
-  async deleteOne(id: number) {
+  async deleteOne(userId: number, id: number) {
     const board = await this.getBoardById(id);
     if (!board) {
       throw new HttpException('NotFound', HttpStatus.NOT_FOUND);
+    }
+    if (board.userId !== userId) {
+      throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
     }
     return this.boardRepository.remove(board);
   }
